GetX: Practice getx, controller, provider, bottom sheet, listview and sheet issue
