+++ flutter_usecases/1-manage-states/practice_syntax_from_0/lib/app/home/binding/home_binding.dart
@@ -1,12 +1,10 @@
-/*
 import 'package:get/get.dart';
-
-import '../controller/home_controller.dart';
+import '../controllers/home_controller.dart';
 
 class HomeBinding extends Bindings {
+
   @override
   void dependencies() {
-    Get.put(HomeController());
+    Get.lazyPut(() => HomeController());
   }
-}*/
-
+}
