+++ flutter_usecases/1-manage-states/practice_syntax_from_0/lib/app/home/providers/home_data_provider.dart
@@ -1,14 +1,13 @@
 import 'package:get/get.dart';
 
-class HomeProvider extends GetConnect {
+class HomeDataProvider extends GetConnect {
 
   Future<List<dynamic>> getHomeData() async {
     final response = await get("https://randomuser.me/api/?results=20");
 
-    if (response.status.hasError) {
+    if(response.status.hasError) {
       return Future.error(response.statusText!);
     } else {
-      print('[---ok---] res_data: ${response.body['results']}');
       return response.body['results'];
     }
   }
