--- flutter_usecases/1-manage-states/practice_syntax_from_0/lib/app/modules/login/controllers/login_controller.dart
@@ -1,28 +1,0 @@
-// import 'package:get/get.dart';
-//
-// class LoginController extends GetxController {
-//   @override
-//   void onInit() {
-//     // TODO: implement onInit
-//     super.onInit();
-//   }
-//
-//   @override
-//   void dispose() {
-//     // TODO: implement dispose
-//     super.dispose();
-//   }
-//
-//   @override
-//   void onClose() {
-//     // TODO: implement onClose
-//     super.onClose();
-//   }
-//
-//   @override
-//   void onReady() {
-//     // TODO: implement onReady
-//     super.onReady();
-//   }
-// }
-
