+++ flutter_usecases/1-manage-states/practice_syntax_from_0/lib/app/routes/app_pages.dart
@@ -1,43 +1,12 @@
-// import 'package:flutter/material.dart';
-// import 'package:get/get.dart';
-//
-// import '../modules/home/binding/home_binding.dart';
-// import '../modules/home/views/home_view.dart';
-// import '../modules/login/views/login_view.dart';
-//
-// part 'app_routes.dart';
-//
-// class AppPages {
-//   AppPages._();
-//
-//   static const INITIAL = Routes.HOME;
-//
-//   static final routes = [
-//     GetPage(
-//         name: _Paths.HOME,
-//         page: () => HomeView(),
-//         binding: HomeBinding(),
-//     ),
-//     GetPage(
-//         name: _Paths.LOGIN,
-//         page: () => LoginView()
-//     ),
-//     GetPage(name: _Paths.SIGNUP, page: () => Container()),
-//   ];
-// }
-
-import 'package:get/get.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import '../home/views/home_view.dart';
+import '../home/binding/home_binding.dart';
+import '../login/views/login_view.dart';
 
-import '../modules/home/binding/home_binding.dart';
-import '../modules/home/views/home_view.dart';
-import '../modules/login/views/login_view.dart';
-
-part 'app_routes.dart';
+part 'app_route.dart';
 
 class AppPages {
-  // AppPages._();
-
   static const INITIAL = Routes.HOME;
 
   static final routes = [
@@ -48,8 +17,8 @@ class AppPages {
     ),
     GetPage(
         name: Routes.LOGIN,
-        page: () => LoginView()
+        page: () => LoginView(),
     ),
-    GetPage(name: Routes.SIGNUP, page: () => Container()),
+    GetPage(name: Routes.SIGNUP, page: () => Container())
   ];
 }
