--- flutter_usecases/1-manage-states/practice_syntax_from_0/lib/app/routes/app_routes.dart
@@ -1,31 +1,0 @@
-// part of 'app_pages.dart';
-
-// abstract class Routes {
-//   Routes_();
-//
-//   static const HOME = _Paths.HOME;
-//   static const LOGIN = _Paths.LOGIN;
-//   static const SIGNUP = _Paths.SIGNUP;
-// }
-
-// abstract class _Paths {
-//   static const HOME = "/home";
-//   static const LOGIN = "/login";
-//   static const SIGNUP = "/signup";
-// }
-
-part of 'app_pages.dart';
-
-abstract class Routes {
-  // Routes_();
-
-  static const HOME = _Paths.HOME;
-  static const LOGIN = _Paths.LOGIN;
-  static const SIGNUP = _Paths.SIGNUP;
-}
-
-abstract class _Paths {
-  static const HOME = "/home";
-  static const LOGIN = "/login";
-  static const SIGNUP = "/signup";
-}
