--- flutter_usecases/1-manage-states/practice_syntax_from_0/lib/providers/UserProvider.dart
@@ -1,15 +1,0 @@
-/*
-import 'package:get/get.dart';
-
-class UserProvider extends GetConnect {
-  Future<List<dynamic>> getUsers() async {
-    final response = await get('https://randomuser.me/api/?results=20');
-
-    if (response.status.hasError) {
-      return Future.error(response.statusText!);
-    } else {
-      return response.body['results'];
-    }
-  }
-}
-*/
