--- flutter_usecases/1-manage-states/practice_syntax_from_0/lib/routes/routes.dart
@@ -1,7 +1,0 @@
-// abstract class Routes {
-//   static String HOME = "/home";
-//   static String LOGIN = "/login";
-//   static String SIGNUP = "/signup";
-//
-//   static String INITIAL = HOME;
-// }
